Added error handling

+++ main.cpp
@@ -37,6 +37,7 @@ using namespace boost::posix_time;
 int main(int argc, char **argv)
 {
     boost::filesystem::create_directory("/input/");
+    boost::filesystem::create_directory("/input/failed/");
     boost::filesystem::create_directory("/output/");
 
     while(true)
@@ -61,23 +62,43 @@ int main(int argc, char **argv)
         size_t count(0);
         for (boost::filesystem::directory_iterator itr(p); itr != end_itr; ++itr)
         {
-            ++count;
-            if (is_regular_file(itr->path()) && itr->path().filename().string().substr(0,1) != ".") {
+            if (is_regular_file(itr->path()) && itr->path().filename().string().substr(0,1) != ".")
+            {
                 std::string current_file = itr->path().string();
                 std::cout << current_file << std::endl;
 
-                pcpp::PcapNgFileReaderDevice pcapReader(current_file.c_str());
-                pcapReader.open();
-
-                pcpp::RawPacket rawPacket;
-                std::string comment;
-                while (pcapReader.getNextPacket(rawPacket, comment)) {
-                    pcapNgWriter.writePacket(rawPacket, comment.c_str());
+                //pcpp::PcapFileReaderDevice pcapReader(current_file.c_str());
+                pcpp::IFileReaderDevice* reader = pcpp::IFileReaderDevice::getReader(current_file.c_str());
+                if(reader && reader->open())
+                {
+                    ++count;
+                    pcpp::PcapNgFileReaderDevice* pcapNgReader(dynamic_cast<pcpp::PcapNgFileReaderDevice*>(reader));
+                    pcpp::PcapFileReaderDevice* pcapReader(dynamic_cast<pcpp::PcapFileReaderDevice*>(reader));
+                    if(pcapNgReader)
+                    {
+                        pcpp::RawPacket rawPacket;
+                        std::string comment;
+                        while (pcapNgReader->getNextPacket(rawPacket, comment)) {
+                            pcapNgWriter.writePacket(rawPacket, comment.c_str());
+                        }
+                    }
+                    else if(pcapReader)
+                    {
+                        pcpp::RawPacket rawPacket;
+                        while (pcapReader->getNextPacket(rawPacket)) {
+                            pcapNgWriter.writePacket(rawPacket);
+                        }
+                    }
+                    boost::filesystem::remove(current_file);
+                    addDelay = false;
+                    reader->close();
                 }
-
-                pcapReader.close();
-                boost::filesystem::remove(current_file);
-                addDelay = false;
+                else
+                {
+                    std::string failedFile("/input/failed/" + itr->path().filename().string());
+                    std::cout << "Failed:" << current_file << " file moved to:" << failedFile << std::endl;
+                    boost::filesystem::rename(current_file, failedFile);
+                }
             }
             if(count == mergeCount)
             {
@@ -85,13 +106,13 @@ int main(int argc, char **argv)
                 break;
             }
         }
-        if(count < mergeCount)
-	{
-		std::cout << "only(" << count << ") file merged" << std::endl;
-            	sleep(1);
-	}
         pcapNgWriter.close();
         boost::filesystem::rename(tmpFile,closeFile);
+        if(count < mergeCount)
+        {
+            std::cout << "only(" << count << ") file merged" << std::endl;
+            sleep(1);
+        }
         std::cout << "closeFile=" << closeFile << std::endl;
         if(addDelay)
         {
